Ensure to run the same perltidy versions as downstream projects
By using our os-autoinst_dev container for running perltidy same as in e.g.
os-autoinst or openQA we are running the same perltidy version as in there.
This fixes the problem that a new release of Perl::Tidy on CPAN would bring in
a need to update to the new version whereas in downstream projects we are
relying on the version that is in openSUSE Tumbleweed and Leap.